+++ embedded-io-adapters/src/futures_03.rs
@@ -57,7 +57,11 @@ impl<T: futures::io::AsyncBufRead + Unpin + ?Sized> embedded_io_async::BufRead f
 
 impl<T: futures::io::AsyncWrite + Unpin + ?Sized> embedded_io_async::Write for FromFutures<T> {
     async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        poll_fn(|cx| Pin::new(&mut self.inner).poll_write(cx, buf)).await
+        match poll_fn(|cx| Pin::new(&mut self.inner).poll_write(cx, buf)).await {
+            Ok(0) if !buf.is_empty() => Err(std::io::ErrorKind::WriteZero.into()),
+            Ok(n) => Ok(n),
+            Err(e) => Err(e),
+        }
     }
 
     async fn flush(&mut self) -> Result<(), Self::Error> {