+++ embedded-io-adapters/src/tokio_1.rs
@@ -68,7 +68,11 @@ impl<T: tokio::io::AsyncBufRead + Unpin + ?Sized> embedded_io_async::BufRead for
 
 impl<T: tokio::io::AsyncWrite + Unpin + ?Sized> embedded_io_async::Write for FromTokio<T> {
     async fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-        poll_fn(|cx| Pin::new(&mut self.inner).poll_write(cx, buf)).await
+        match poll_fn(|cx| Pin::new(&mut self.inner).poll_write(cx, buf)).await {
+            Ok(0) if !buf.is_empty() => Err(std::io::ErrorKind::WriteZero.into()),
+            Ok(n) => Ok(n),
+            Err(e) => Err(e),
+        }
     }
 
     async fn flush(&mut self) -> Result<(), Self::Error> {