+++ embedded-io-adapters/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add support for adapting `BufRead` from `futures` and `tokio`.
+- Return an error when a wrapped `std`/`futures`/`tokio` `write()` call returns
+  `Ok(0)` to comply with `embedded_io::Write` requirements.
 
 ## 0.5.0 - 2023-08-06